+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/mapper/HighPassExistenceMapperRegistry.java
@@ -24,6 +24,7 @@ import java.util.Objects;
  * @author mooyuan
  * @since 2.0.0
  */
+@Component
 public class HighPassExistenceMapperRegistry extends AbstractMapperRegistry {
 
     public static final String MAPPER_TYPE = "high_pass_existence_mapper";