+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/mapper/LowPassCounterMapperRegistry.java
@@ -24,6 +24,7 @@ import java.util.List;
  * @author mooyuan
  * @since 2.0.0
  */
+@Component
 public class LowPassCounterMapperRegistry extends AbstractMapperRegistry {
 
     public static final String MAPPER_TYPE = "low_pass_counter_mapper";